requirements.txt
asttokens==3.0.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.1.0
pytz==2024.2
setuptools==70.0.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
torch==2.5.1+rocm6.2
torchaudio==2.5.1+rocm6.2
torchsummary==1.5.1
torchvision==0.20.1+rocm6.2
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
wcwidth==0.2.13
